pyproject.toml
[project]
name = "mcp-server-email"
version = "0.1.0"
description = "Give your AI assistant an email address"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "google-api-python-client>=2.166.0",
 "google-auth>=2.38.0",
 "google-auth-oauthlib>=1.2.1",
 "mcp>=1.6.0",
 "pydantic>=2.11.3",
 "python-dotenv>=1.1.0",
]
[[project.authors]]
name = "R.Keelan"
email = "R.Keelan@gmail.com"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-email = "mcp_server_email:main"
